# SOLR-16611: fix collapse hint=top_fc bug