--- a/solr/core/src/test/org/apache/solr/search/TestCollapseQParserPlugin.java
+++ b/solr/core/src/test/org/apache/solr/search/TestCollapseQParserPlugin.java
@@ -443,9 +443,7 @@ public class TestCollapseQParserPlugin extends SolrTestCaseJ4 {
     ModifiableSolrParams params = new ModifiableSolrParams();
     params.add("q", "*:*");
     params.add("fq", "{!collapse field=" + group + "" + hint + "}");
-    assertQ(
-            req(params, "indent", "on"),
-            "*[count(//doc)=0]");
+    assertQ(req(params, "indent", "on"), "*[count(//doc)=0]");
 
     String[] doc = {
       "id", "1", "term_s", "YYYY", group, "1", "test_i", "5", "test_l", "10", "test_f", "2000"
@@ -1101,7 +1099,7 @@ public class TestCollapseQParserPlugin extends SolrTestCaseJ4 {
             " sort='bogus_sort_s desc' ",
           }) {
 
-        ModifiableSolrParams params = null;
+        ModifiableSolrParams params;
 
         // w/default nullPolicy, no groups found
         params = new ModifiableSolrParams();